+++ kipper/core/src/compiler/analysis/analyser/semantic-checker.ts
@@ -37,6 +37,7 @@ import {
 	UndefinedReferenceError,
 	UnknownReferenceError,
 	InvalidJumpStatementError,
+	BuiltInOrInternalGeneratorFunctionNotFoundError,
 } from "../../../errors";
 
 /**
@@ -61,7 +62,8 @@ export class KipperSemanticChecker extends KipperSemanticsAsserter {
 				? scopeCtx.innerScope.getEntryRecursively(identifier)
 				: this.programCtx.globalScope.getEntry(identifier)) ??
 			this.programCtx.globalScope.getEntry(identifier) ?? // Fall back to looking globally
-			this.programCtx.getBuiltInFunction(identifier) // Fall back to searching through built-in functions
+			this.programCtx.getBuiltInFunction(identifier) ?? // Fall back to searching through built-in functions
+			this.programCtx.getBuiltInVariable(identifier) // Fall back to searching through built-in variables
 		);
 	}
 
@@ -140,11 +142,28 @@ export class KipperSemanticChecker extends KipperSemanticsAsserter {
 	 */
 	public globalCanBeRegistered(identifier: string): void {
 		let identifierAlreadyExists: boolean = this.programCtx.globalScope.getEntry(identifier) !== undefined;
-		let globalAlreadyExists: boolean = this.programCtx.getBuiltInFunction(identifier) !== undefined;
+		let globalAlreadyExists: boolean =
+			this.programCtx.getBuiltInFunction(identifier) !== undefined ||
+			this.programCtx.getBuiltInVariable(identifier) !== undefined;
 
 		// If the identifier is already used or the global already exists, throw an error
 		if (identifierAlreadyExists || globalAlreadyExists) {
 			throw this.assertError(new InvalidGlobalError(identifier));
+		}
+	}
+
+	/**
+	 * Asserts that the passed identifier is a valid built-in global that has a generator function in the
+	 * {@link this.programCtx.target.builtInGenerator BuiltInGenerator} of the target.
+	 * @param identifier The identifier to check.
+	 * @throws {BuiltInOrInternalGeneratorFunctionNotFoundError} If there is no generator function for the passed
+	 * identifier.
+	 * @since 0.10.0
+	 */
+	public globalCanBeGenerated(identifier: string): void {
+		const generator = Reflect.get(this.programCtx.target.builtInGenerator, identifier);
+		if (generator === undefined) {
+			throw this.assertError(new BuiltInOrInternalGeneratorFunctionNotFoundError(identifier));
 		}
 	}