+++ kipper/target-js/src/target.ts
@@ -4,7 +4,7 @@
  * @copyright 2021-2022 Luna Klatzer
  * @since 0.10.0
  */
-import { KipperCompileTarget } from "@kipper/core";
+import { BuiltInFunction, BuiltInVariable, KipperCompileTarget } from "@kipper/core";
 import { JavaScriptTargetSemanticAnalyser } from "./semantic-analyser";
 import { JavaScriptTargetCodeGenerator } from "./code-generator";
 import { JavaScriptTargetBuiltInGenerator } from "./built-in-generator";
@@ -93,6 +93,23 @@ export class KipperJavaScriptTarget extends KipperCompileTarget {
 		builtInGenerator: JavaScriptTargetBuiltInGenerator = new JavaScriptTargetBuiltInGenerator(),
 	) {
 		super("javascript", semanticAnalyser, codeGenerator, builtInGenerator, "js");
+	}
+
+	/**
+	 * Fetches the reserved identifier for the translated code.
+	 *
+	 * This will also ensure that {@link BuiltInVariable.local local variables} are not registered onto the global object.
+	 * Those will simply stay as local variables with the same identifier.
+	 * @param signature The identifier or signature object to translate to its JavaScript form.
+	 * @since 0.10.0
+	 */
+	public static getBuiltInIdentifier(signature: string | BuiltInVariable | BuiltInFunction): string {
+		if (typeof signature === "string") {
+			return `${this.internalObjectIdentifier}.${signature}`;
+		}
+		return "local" in signature && signature.local
+			? signature.identifier
+			: this.getBuiltInIdentifier(signature.identifier);
 	}
 }